+++ src/Models/calendar.py
@@ -16,7 +16,7 @@ class Calendar(Base):
     sunday = Column(Boolean, nullable=False)
     start_date = Column(Numeric, nullable=True)
     end_date = Column(Numeric, nullable=False)
-    calendar_dates_hash = Column(BigInteger, nullable=True)
+    calendar_dates_hash = Column(Numeric, nullable=True)
 
     def __repr__(self):
         return "<Calendar(id='{}', start='{}', end={})>" \
